Split v4/5 and add websocket imports
