+++ plugins/importer-insomnia/tests/fixtures/version-5.output.json
@@ -5,6 +5,7 @@
         "createdAt": "2025-01-13T15:15:43.767",
         "updatedAt": "2025-01-13T15:15:55.209",
         "base": true,
+        "public": true,
         "id": "GENERATE_ID::env_20945044d3c8497ca8b717bef750987e",
         "model": "environment",
         "name": "Base Environment",
@@ -21,6 +22,7 @@
         "createdAt": "2025-01-13T15:15:58.515",
         "updatedAt": "2025-01-13T15:16:34.705",
         "base": false,
+        "public": true,
         "id": "GENERATE_ID::env_6f7728bb7fc04d558d668e954d756ea2",
         "model": "environment",
         "name": "Production",
@@ -38,6 +40,7 @@
         "createdAt": "2025-01-13T15:16:14.707",
         "updatedAt": "2025-01-13T15:16:31.078",
         "base": false,
+        "public": true,
         "id": "GENERATE_ID::env_976a8b6eb5d44fb6a20150f65c32d243",
         "model": "environment",
         "name": "Staging",
@@ -60,7 +63,7 @@
         "id": "GENERATE_ID::fld_42eb2e2bb22b4cedacbd3d057634e80c",
         "model": "folder",
         "name": "Top Level",
-        "sortPriority": 0,
+        "sortPriority": -1736781404718,
         "workspaceId": "GENERATE_ID::wrk_c1eacfa750a04f3ea9985ef28043fa53"
       }
     ],
@@ -75,7 +78,7 @@
         "method": "SayHello",
         "name": "New Request",
         "service": "hello.HelloService",
-        "sortPriority": 1,
+        "sortPriority": -1746799344864,
         "updatedAt": "2025-05-09T14:05:44.082",
         "url": "grpcb.in:9000",
         "workspaceId": "GENERATE_ID::wrk_c1eacfa750a04f3ea9985ef28043fa53"
@@ -124,8 +127,34 @@
         "method": "GET",
         "model": "http_request",
         "name": "New Request",
-        "sortPriority": 0,
+        "sortPriority": -1736781406672,
         "url": "${[BASE_URL ]}/foo/:id",
+        "workspaceId": "GENERATE_ID::wrk_c1eacfa750a04f3ea9985ef28043fa53"
+      }
+    ],
+    "websocketRequests": [
+      {
+        "id": "GENERATE_ID::ws-req_5d1a4c7c79494743962e5176f6add270",
+        "createdAt": "2025-05-09T14:05:53.909",
+        "updatedAt": "2025-05-10T14:25:20.958",
+        "message": "",
+        "model": "websocket_request",
+        "name": "New WebSocket Request",
+        "sortPriority": -1746799553909,
+        "authenticationType": "basic",
+        "authentication": {
+          "password": "password",
+          "username": "user"
+        },
+        "folderId": null,
+        "headers": [
+          {
+            "enabled": true,
+            "name": "User-Agent",
+            "value": "insomnia/11.1.0"
+          }
+        ],
+        "url": "wss://echo.websocket.org",
         "workspaceId": "GENERATE_ID::wrk_c1eacfa750a04f3ea9985ef28043fa53"
       }
     ],
